api/requirements.txt
Flask==3.0.3
psycopg[binary]==3.2.1
python-dotenv==1.0.1
